refactor(create-stylable-app): organize files in dedicated directories, use st-import and use new reset.st.css
