+++ packages/create-stylable-app/template/ts-react-rollup/src/components/app.tsx
