+++ packages/create-stylable-app/template/ts-react-rollup/src/components/header.st.css
