+++ packages/create-stylable-app/template/ts-react-rollup/src/index.tsx
@@ -1,6 +1,7 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import { App } from './app';
+import { App } from './components/app';
+import './globals.st.css';
 import './reset.st.css';
 
 ReactDOM.render(
