--- packages/create-stylable-app/template/ts-react-rollup/src/reset.st.css
@@ -1,45 +1,0 @@
-:global(*),
-:global(::before),
-:global(::after) {
-    box-sizing: border-box;
-}
-
-:global(html),
-:global(body) {
-    margin: 0;
-    padding: 0;
-}
-
-:global(html) {
-    font-family:
-        system-ui,
-        /* macOS 10.11-10.12 */
-        -apple-system,
-        /* Windows 6+ */
-        "Segoe UI",
-        /* Android 4+ */
-        "Roboto",
-        /* Ubuntu 10.10+ */
-        "Ubuntu",
-        /* Gnome 3+ */
-        "Cantarell",
-        /* KDE Plasma 5+ */
-        "Noto Sans",
-        /* fallback */
-        sans-serif,
-        /* macOS emoji */
-        "Apple Color Emoji",
-        /* Windows emoji */
-        "Segoe UI Emoji",
-        /* Windows emoji */
-        "Segoe UI Symbol",
-        /* Linux emoji */
-        "Noto Color Emoji";
-    ;
-}
-
-:global(a),
-:global(a:visited) {
-    color: inherit;
-    cursor: pointer;
-}
