+++ packages/create-stylable-app/template/ts-react-webpack/src/components/app.tsx
