+++ packages/create-stylable-app/template/ts-react-webpack/src/styles/globals.st.css
@@ -1,0 +1,4 @@
+:global(html) {
+  background-color: rgb(37, 37, 38);
+  color: white;
+}
